Making the code more reusable
The MIDI slider this was created for is fantastic as a general purpose input device. However, David had a specific use in mind. I've stripped out much of his code and just have it place the data from sliders and buttons into some variables. (Notice: This is my first time using Python )